add FreeCAD file with HDPE drum and expanded metal mesh
This commit adds our base FreeCAD file 'intake-drum.FCStd', which we will be attempting to optimize in subsequent commits.

This file effectively only has two objects:

 1. A HDPE drum, cut in half length-wise

 2. An expanded metal mesh, which rests atop the half-drum

Despite its relative simplicity, the (compressed) FCStd file is 13 MB.

Uncompressed, the object is made-up of the following layers:

-------------------------------
MemSize (bytes) | Object Label
-------------------------------
    108,688,850 | intake-drum
    108,638,852 | expanded-metal-mesh
         48,864 | intake-drum_thickness
            149 | intake drum
            105 | XY_Plane
            105 | XZ_Plane
            105 | YZ_Plane
            101 | X_Axis
            101 | Y_Axis
            101 | Z_Axis
             56 | Origin
-------------------------------

From above, we can see the entire file's (uncompressed) MemSize is about 104 MB (108,688,850/1,024/1,024). Of that, 99.95% is just the "expanded-metal-mesh" object (108,638,852/108,688,850*100)

TODO: in subsequent commits, we'll add new files that attempt to reduce the size of this FreeCAD file by optimizing the expanded-metal-mesh object