add different sized Expanded Metal sheets
This commit adds 4 new FreeCAD files, all of which were copied from 'expanded-metal-array.fcstd1'

Each of the files has a very flat "cube" of a different size, and I just incremented the array of each such that the Expanded Metal array object *just* covered the entire surface area of the cube. There's 4 sizes:

 1. 10 cm x 10 cm
 2. 20 cm x 20 cm
 3. 40 cm x 40 cm
 4. 80 cm x 80 cm

I also tried to create one 160 cm x 160 cm, but it crashed FreeCAD.

And before that, I first thought I'd do 1 meter, 10 meter, and 100 meters, but that definitely didn't work; the app caused my computer to thrash and I had to kill FreeCAD.

Once I created each of these files, I recorded the (compressed) filesize of each FreeCAD FCStd file, the MemSize of the total document, and the MemSize of the highest array in the tree (for just the 'expanded-metal-array' object). Then I put this data into a simple spreadsheet and created a chart in libreoffice (expanded-metal-array.ods).

The chart in this file appears to show an exponential growth in the MemSize of the FreeCAD file, even when plotting the data linearly.

Also of note is that the actual on-disk compressed .FCStd file size increases negligibly.

This shows that you can't rely on the actual file size of the .FCStd file to determine if your design is sane. You have to open the app and calculate the MemSize of the objects to know if it's going to scale.

And now we know that simplifying an object like an Expanded Metal Mesh into a very simple sketch and then just "copying" that with a series of arrays is *not* a solution that scales.