add different sized Expanded Metal sheets (hatch)
This commit adds 7 new FreeCAD files

Each of the files has a very flat "cube" of different sizes:

  1. 10 cm x 10 cm
  2. 20 cm x 20 cm
  3. 40 cm x 40 cm
  4. 80 cm x 80 cm
  5. 160 cm x 160 cm
  6. 320 cm x 320 cm
  7. 640 cm x 640 cm

I also tried to create a "1,280 cm x 1,280 cm" file, but it crashed FreeCAD. Already, though, we can see that the addition of these 3 doubled files over the last commit shows the optimization improvement of "hatch approach" vs the "array approach" (where it would crash if I had tried a 160 cm x 160 cm sheet)

As before, I also created a spreadsheet that shows the (compressed) filesize of each of these FreeCAD FCStd files, and the (uncompressed) MemSize of the total document (expanded-metal-hatch.ods).

The chart shows a linear growth. And the MemSize is *much* smaller than the previous chart, on the order of three magnitudes (eg the 80cm squared MemSize was 0.78 MB on the "hatch" file vs 165 MB on the "array" file).

The fact that we couldn't make a sheet of metal that's >10 meters shows that this "hatch" solution still has its limitations, but it's clearly a better option than the "array" solution