added facebinder
This commit updates our "intake drum with a hatch" to use a facebinder
object between the cube and the hatch. This makes the hatch apply only
to the top surface of the cube, so it looks more like a sheet of
expanded metal, rather than a square frame that's been wrapped on
all-sides by a piece of wire

 * https://forum.freecad.org/viewtopic.php?p=846124

Curiously, the (compressed) size of the file is no 20KB and the total
(uncompressed) MemSize of the document is 104,992 bytes. That's 0.020 MB
& 0.100 MB, respectfully -- which is smaller than I listed in my last
commit (0.32 MB & 0.21 MB). So adding a facebinder as an additional
object *decreased* the size? Huh.